refactor: relocate agent control-plane clients into conductor-client
AgentClient (compile/deploy/start/status/respond) and the SSE streaming
client SseClient move from conductor-ai's internal package to
io.orkes.conductor.client; OrkesClients gains getAgentClient(). Their
transport DTOs (AgentRequest, StartResponse, AgentStatusResponse,
PendingTool, RespondBody, CompileResponse) land in
io.orkes.conductor.client.model.agent, and the typed exceptions
(AgentspanException, AgentAPIException, AgentNotFoundException) in
io.orkes.conductor.client.exceptions — the 404 -> AgentNotFoundException
mapping stays on the client (SuiteHttpApi404 contract).

Boundary adaptations, since conductor-client cannot see conductor-ai
types:
- AgentRequest is now a pure JSON transport DTO (@JsonInclude NON_NULL,
  no custom serializer). Domain serialization stays in conductor-ai:
  AgentRuntime.agentRequest() calls AgentConfigSerializer.serialize()
  and resolves the Framework wire name at build time; static plans are
  passed as plan.toJson().
- SseClient is transport-only and emits raw parsed JSON maps ([DONE]
  and type=done handling unchanged); AgentStream maps them to AgentEvent,
  keeping the log-and-skip contract for unmappable events.
- Credential* exceptions stay in conductor-ai, still extending the
  (relocated) AgentspanException base.

Tests for the moved exceptions relocate to conductor-client
(AgentExceptionsTest); docs/agents and CHANGELOG updated (migration is
a coordinate swap plus the relocated imports where used directly).

Validation: spotlessCheck (incl. -Pe2e) green; root test green
(conductor-client 987 incl. +3 relocated, conductor-ai 271/271 with
live server incl. 10 schedule integration tests, 0 failures);
live e2e vs agentspan-server-0.4.0: 138 tests, 137 executed, 0 failed
(1 env-gated media skip, same as baseline); Example01 smoke COMPLETED;
javadoc builds with no new warnings.
