+++ conductor-ai-e2e/src/test/java/Suite10CodeExecution.java
@@ -22,8 +22,9 @@ import org.conductoross.conductor.ai.enums.AgentStatus;
 import org.conductoross.conductor.ai.execution.DockerCodeExecutor;
 import org.conductoross.conductor.ai.execution.ExecutionResult;
 import org.conductoross.conductor.ai.model.AgentResult;
-import org.conductoross.conductor.ai.model.CompileResponse;
 import org.junit.jupiter.api.*;
+
+import io.orkes.conductor.client.model.agent.CompileResponse;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.junit.jupiter.api.Assumptions.assumeTrue;
