+++ conductor-ai-e2e/src/test/java/Suite13Callbacks.java
@@ -21,8 +21,9 @@ import org.conductoross.conductor.ai.AgentRuntime;
 import org.conductoross.conductor.ai.CallbackHandler;
 import org.conductoross.conductor.ai.enums.AgentStatus;
 import org.conductoross.conductor.ai.model.AgentResult;
-import org.conductoross.conductor.ai.model.CompileResponse;
 import org.junit.jupiter.api.*;
+
+import io.orkes.conductor.client.model.agent.CompileResponse;
 
 import static org.junit.jupiter.api.Assertions.*;
