+++ conductor-client/src/main/java/io/orkes/conductor/client/OrkesClients.java
@@ -30,6 +30,10 @@ public class OrkesClients {
         return new OrkesWorkflowClient(client);
     }
 
+    public AgentClient getAgentClient() {
+        return new AgentClient(client);
+    }
+
     public AuthorizationClient getAuthorizationClient() {
         return new OrkesAuthorizationClient(client);
     }
