+++ conductor-client/src/main/java/io/orkes/conductor/client/exceptions/AgentAPIException.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.conductoross.conductor.ai.exceptions;
+package io.orkes.conductor.client.exceptions;
 
 /**
  * Thrown when the Agentspan server returns a non-2xx HTTP response.
