+++ conductor-client/src/main/java/io/orkes/conductor/client/exceptions/AgentNotFoundException.java
@@ -10,18 +10,17 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.conductoross.conductor.ai.exceptions;
+package io.orkes.conductor.client.exceptions;
 
 /**
- * Base exception for all Agentspan SDK errors.
+ * Thrown when the workflow / agent / execution ID is not found (HTTP 404).
+ *
+ * <p>Subclass of {@link AgentAPIException} — catch this when you want to
+ * distinguish "doesn't exist" from generic server errors. Mirrors
+ * {@code AgentNotFoundError} in the Python SDK.
  */
-public class AgentspanException extends RuntimeException {
-
-    public AgentspanException(String message) {
-        super(message);
-    }
-
-    public AgentspanException(String message, Throwable cause) {
-        super(message, cause);
+public class AgentNotFoundException extends AgentAPIException {
+    public AgentNotFoundException(int statusCode, String responseBody) {
+        super(statusCode, responseBody);
     }
 }
