+++ conductor-client/src/main/java/io/orkes/conductor/client/model/agent/RespondBody.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.conductoross.conductor.ai.internal;
+package io.orkes.conductor.client.model.agent;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
